Move file from Raw into Jupyter folder
