+++ src/test/datascience/notebook/remoteNotebookEditor.vscode.common.ts
@@ -28,7 +28,7 @@ import {
 import { openNotebook } from '../helpers';
 import { PYTHON_LANGUAGE, Settings } from '../../../platform/common/constants';
 import { IS_REMOTE_NATIVE_TEST, JVSC_EXTENSION_ID_FOR_TESTS } from '../../constants';
-import { PreferredRemoteKernelIdProvider } from '../../../kernels/raw/finder/preferredRemoteKernelIdProvider';
+import { PreferredRemoteKernelIdProvider } from '../../../kernels/jupyter/preferredRemoteKernelIdProvider';
 import { INotebookControllerManager } from '../../../notebooks/types';
 import { IServiceContainer } from '../../../platform/ioc/types';
